pyproject.toml
[project]
name = "PyCOM"
version = "2.0.0"
description = "A serial port communication tool written with python."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chardet>=5.2.0",
    "pyserial>=3.5",
    "pyside6>=6.10.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
